BAU: Merge utils class into helper class
- One of the methods on the utils class technically wasn't stateless (requiring an OidcAPI instance). I've merged the utils code into the helper class (including the tests)
